refactor(web): move consent resolution/persistence into lib/consent (#38242)
