+++ clients/web/src/domains/onboarding/prechat.ts
@@ -13,7 +13,7 @@
  *
  * Storage-error handling matches the pattern in
  * `@/domains/onboarding/prefs` (e.g. `readSelectedVersion`) and
- * `@/utils/onboarding-cleanup`: every read/write is wrapped in `try/catch` so a
+ * `@/lib/consent/consent-persistence`: every read/write is wrapped in `try/catch` so a
  * disabled or quota-exceeded `sessionStorage` degrades to "no pending
  * context" instead of throwing into the caller.
  */
